start common users page
