+++ cont3xt/vueapp/src/components/pages/Users.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="container-fluid">
+    <UsersCommon />
+  </div>
+</template>
+
+<script>
+import UsersCommon from '../../../../../common/Users';
+
+export default {
+  name: 'Users',
+  components: { UsersCommon }
+};
+</script>
